CIV-3760: Modified code for Additional fee callback handlers
Modified to stop calls to  fee Apis for "additional fee reference" and for obtaining "additional fee values" during camunda  "judicial decesion flow".
